+++ sdk/python/lib/test/langhost/first_class_provider_unknown/__main__.py
@@ -11,20 +11,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from os import path
-from ..util import LanghostTest
+from pulumi import CustomResource, ProviderResource, ResourceOptions, Output
 
+class MyResource(CustomResource):
+    value: Output[int]
 
-class OneResourceTest(LanghostTest):
-    def test_one_resource(self):
-        self.run_test(
-            program=path.join(self.base_path(), "one_resource"),
-            expected_resource_count=1)
+    def __init__(self, name, value, opts=None):
+        CustomResource.__init__(self, "test:index:MyResource", name, props={
+            "value": value,
+        }, opts=opts)
 
-    def register_resource(self, _ctx, _dry_run, ty, name, _resource,
-                          _dependencies, _parent, _custom, _protect, _provider):
-        self.assertEqual(ty, "test:index:MyResource")
-        self.assertEqual(name, "testResource1")
-        return {
-            "urn": self.make_urn(ty, name),
-        }
+class MyProvider(ProviderResource):
+    def __init__(self, name, opts=None):
+        ProviderResource.__init__(self, "test", name, {}, opts)
+
+prov = MyProvider("testprov")
+res = MyResource("res", 42, ResourceOptions(provider=prov))